feat(cosmwasm-vm): refactors and improvements
