+++ vm-wasmi/src/error.rs
@@ -1,0 +1,69 @@
+use core::{
+    fmt::{Debug, Display},
+    num::TryFromIntError,
+};
+use cosmwasm_vm::{
+    executor::ExecutorError,
+    memory::{MemoryReadError, MemoryWriteError},
+    system::SystemError,
+};
+
+#[derive(PartialEq, Eq, Debug)]
+#[allow(clippy::module_name_repetitions)]
+pub enum WasmiVMError {
+    ExecutorError(ExecutorError),
+    SystemError(SystemError),
+    MemoryReadError(MemoryReadError),
+    MemoryWriteError(MemoryWriteError),
+    MemoryNotExported,
+    MemoryExportedIsNotMemory,
+    LowLevelMemoryReadError,
+    LowLevelMemoryWriteError,
+    InvalidPointer,
+    UnexpectedUnit,
+    UnexpectedReturnType,
+    ExpectedUnit,
+    ExpectedPointer,
+    InvalidHostSignature,
+    InvalidValue,
+    MaxLimitExceeded,
+    NotADynamicModule,
+    FunctionNotFound,
+    InternalWasmiError,
+}
+
+impl From<ExecutorError> for WasmiVMError {
+    fn from(e: ExecutorError) -> Self {
+        WasmiVMError::ExecutorError(e)
+    }
+}
+
+impl From<MemoryReadError> for WasmiVMError {
+    fn from(e: MemoryReadError) -> Self {
+        WasmiVMError::MemoryReadError(e)
+    }
+}
+
+impl From<MemoryWriteError> for WasmiVMError {
+    fn from(e: MemoryWriteError) -> Self {
+        WasmiVMError::MemoryWriteError(e)
+    }
+}
+
+impl From<SystemError> for WasmiVMError {
+    fn from(e: SystemError) -> Self {
+        WasmiVMError::SystemError(e)
+    }
+}
+
+impl From<TryFromIntError> for WasmiVMError {
+    fn from(_: TryFromIntError) -> Self {
+        WasmiVMError::InvalidPointer
+    }
+}
+
+impl Display for WasmiVMError {
+    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
+        write!(f, "{self:?}")
+    }
+}
